added files to repository

+++ GameOfShapes/enemy.cpp
@@ -55,17 +55,19 @@ SDL_Rect Enemy::setCollisionBox(int x, int y, int width, int height) const {
 vector<Vector2D> Enemy::vertices() const {
     vector<Vector2D> enemyVertices;
     
-    float vertex1X = transform.position.x;
-    float vertex1Y = transform.position.y + 444;
-    float vertex2X = transform.position.x + 512 / 2;
-    float vertex2Y = transform.position.y;
-    float vertex3X = transform.position.x + 512;
-    float vertex3Y = transform.position.y + 444;
+    float vertex1X = (float) transform.position.x;
+    float vertex1Y = (float) (transform.position.y + 444);
+    float vertex2X = (float) (transform.position.x + 512 / 2);
+    float vertex2Y = (float) transform.position.y;
+    float vertex3X = (float) (transform.position.x + 512);
+    float vertex3Y = (float) (transform.position.y + 444);
     
     enemyVertices.push_back(Vector2D(vertex1X, vertex1Y));
     enemyVertices.push_back(Vector2D(vertex2X, vertex2Y));
     enemyVertices.push_back(Vector2D(vertex3X, vertex3Y));
     
+    //for debug
+    //cout << enemyVertices[1].x << ", " << enemyVertices[1].y << endl;
     return enemyVertices;
 }
